stream: move docker tests into stream_docker_test.go
adds TestStreamDial and TestStreamListen, which fail because they reference an uninitialized logger
